requirements.txt
semantic-kernel==0.9.0
langchain==0.1.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.1.2
PyPDF2==3.0.1
tqdm==4.66.1
python-dotenv==1.0.0
nest-asyncio==1.5.8